requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
numpy==1.26.4; python_version < "3.12"
numpy>=2.0.0,<2.2; python_version >= "3.12"
faiss-cpu>=1.12.0
python-docx==1.1.0
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
google-generativeai==0.7.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.2.0
tqdm==4.66.1
typing_extensions==4.10.0
PyMuPDF==1.23.8; python_version < "3.12"
PyMuPDF>=1.24.0; python_version >= "3.12"
pandas==2.2.2
matplotlib==3.8.4